change to parsing hardcoded JSON data and small change to the UI

+++ app/src/main/java/com/example/android/quakereport/Earthquake.java
@@ -6,17 +6,17 @@ package com.example.android.quakereport;
 
 public class Earthquake {
 
-    private float mMagitude;
+    private double mMagitude;
     private String mPlace;
     private String mTime;
 
-    public Earthquake (float Magitude, String Place, String Time){
+    public Earthquake (double Magitude, String Place, String Time){
         mMagitude = Magitude;
         mPlace = Place;
         mTime = Time;
     }
 
-    public float getMagitude() { return mMagitude;}
+    public double getMagitude() { return mMagitude;}
     public String getPlace() {return  mPlace;}
     public String getTime() {return mTime;}
 }
